refactor: moved BasicInfoForm to Profile/ root
