+++ client/src/pages/Profile.jsx
@@ -2,7 +2,7 @@ import { Typography } from '@mui/material';
 import React from 'react';
 
 import AvatarUpload from '../components/Profile/AvatarUpload';
-import BasicInfoForm from '../components/Profile/Forms/BasicInfoForm';
+import BasicInfoForm from '../components/Profile/BasicInfoForm';
 
 export default function Profile() {
   return (
